cache invalidation feature is complete
